# Conflicts:
#	dashboards-ui/dashboardsui-web/public_html/@version@/js/builder/right.panel.model.js
#	dashboards-ui/dashboardsui-web/public_html/builder.html

+++ dashboards-entities/src/integTest/java/oracle/sysman/emaas/platform/dashboards/entity/EmsDashboardSetPKTest.java
@@ -1,8 +1,8 @@
 package oracle.sysman.emaas.platform.dashboards.entity;
 
-import org.testng.annotations.Test;
+import java.math.BigInteger;
 
-import static org.testng.Assert.*;
+import org.testng.annotations.Test;
 
 /**
  * Created by xiadai on 2016/11/1.
@@ -12,20 +12,20 @@ public class EmsDashboardSetPKTest {
     private EmsDashboardSetPK emsDashboardSetPK = new EmsDashboardSetPK();
     @Test
     public void testGetDashboardSetId() throws Exception {
-        emsDashboardSetPK.setDashboardSetId(1L);
+        emsDashboardSetPK.setDashboardSetId(BigInteger.valueOf(1L));
         emsDashboardSetPK.getDashboardSetId();
     }
 
     @Test
     public void testGetSubDashboardId() throws Exception {
-        emsDashboardSetPK.setSubDashboardId(1L);
+        emsDashboardSetPK.setSubDashboardId(BigInteger.valueOf(1L));
         emsDashboardSetPK.getSubDashboardId();
     }
 
     @Test
     public void testEquals() throws Exception {
         EmsDashboardSetPK emsDashboardSetPKIns = new EmsDashboardSetPK();
-        emsDashboardSetPKIns.setDashboardSetId(1L);
+        emsDashboardSetPKIns.setDashboardSetId(BigInteger.valueOf(1L));
         emsDashboardSetPK.equals(emsDashboardSetPKIns);
         emsDashboardSetPK.equals(new EmsDashboard());
     }
